Remove type conversion

+++ sksurv/tree/tree.py
@@ -244,7 +244,7 @@ class SurvivalTree(BaseEstimator, SurvivalAnalysisMixin):
         # Build tree
         self.criterion = "logrank"
         criterion = LogrankCriterion(
-            self.n_outputs_, n_samples, self.unique_times_, self.is_event_time_.astype(np.int32)
+            self.n_outputs_, n_samples, self.unique_times_, self.is_event_time_
         )
 
         SPLITTERS = SPARSE_SPLITTERS if issparse(X) else DENSE_SPLITTERS
